force bootstrap classloader in primative TypeReferences
FieldReference instances instantiated for containerised classes would
always refer to the classloader of the containing class. When the
FieldReference is to a primative type this causes issues in the baseline
compiler (particularly checks relying in is"Type"() methods. Atoms can
now be checked to see if they describe primatives and the appropriate
classloader is chosen in TypeReference.findOrCreate.